Split parsing into separate crate
To be able to use it for Python nodes in the future.
