+++ msg-gen-macro/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "dora-ros2-bridge-msg-gen"
+name = "dora-ros2-bridge-msg-gen-macro"
 version = "0.1.0"
 edition = "2021"
 authors = ["Yuma Hiramatsu <yuma.hiramatsu@gmail.com>"]
@@ -16,3 +16,7 @@ quote = "1.0"
 regex = "1"
 syn = "1.0"
 thiserror = "1.0"
+dora-ros2-bridge-msg-gen = { path = "../msg-gen" }
+
+[lib]
+proc-macro = true
