+++ msg-gen-macro/test_msgs/action/Fibonacci.action
@@ -1,0 +1,8 @@
+#goal definition
+int32 order
+---
+#result definition
+int32[] sequence
+---
+#feedback
+int32[] sequence
