+++ msg-gen-macro/test_msgs/msg/Constants.msg
@@ -1,0 +1,13 @@
+bool BOOL_CONST=true
+byte BYTE_CONST=50
+char CHAR_CONST=100
+float32 FLOAT32_CONST=1.125
+float64 FLOAT64_CONST=1.125
+int8 INT8_CONST=-50
+uint8 UINT8_CONST=200
+int16 INT16_CONST=-1000
+uint16 UINT16_CONST=2000
+int32 INT32_CONST=-30000
+uint32 UINT32_CONST=60000
+int64 INT64_CONST=-40000000
+uint64 UINT64_CONST=50000000
