+++ msg-gen-macro/test_msgs/msg/MultiNested.msg
@@ -1,0 +1,10 @@
+# Mulitple levels of nested messages
+Arrays[3] array_of_arrays
+BoundedSequences[3] array_of_bounded_sequences
+UnboundedSequences[3] array_of_unbounded_sequences
+Arrays[<=3] bounded_sequence_of_arrays
+BoundedSequences[<=3] bounded_sequence_of_bounded_sequences
+UnboundedSequences[<=3] bounded_sequence_of_unbounded_sequences
+Arrays[] unbounded_sequence_of_arrays
+BoundedSequences[] unbounded_sequence_of_bounded_sequences
+UnboundedSequences[] unbounded_sequence_of_unbounded_sequences
