+++ msg-gen-macro/test_msgs/msg/Nested.msg
@@ -1,0 +1,1 @@
+BasicTypes basic_types_value
