+++ msg-gen-macro/test_msgs/msg/Strings.msg
@@ -1,0 +1,13 @@
+string string_value
+string string_value_default1 "Hello world!"
+string string_value_default2 "Hello'world!"
+string string_value_default3 'Hello"world!'
+string string_value_default4 'Hello\'world!'
+string string_value_default5 "Hello\"world!"
+string STRING_CONST="Hello world!"
+string<=22 bounded_string_value
+string<=22 bounded_string_value_default1 "Hello world!"
+string<=22 bounded_string_value_default2 "Hello'world!"
+string<=22 bounded_string_value_default3 'Hello"world!'
+string<=22 bounded_string_value_default4 'Hello\'world!'
+string<=22 bounded_string_value_default5 "Hello\"world!"
