+++ msg-gen-macro/test_msgs/srv/BasicTypes.srv
@@ -1,0 +1,29 @@
+bool bool_value
+byte byte_value
+char char_value
+float32 float32_value
+float64 float64_value
+int8 int8_value
+uint8 uint8_value
+int16 int16_value
+uint16 uint16_value
+int32 int32_value
+uint32 uint32_value
+int64 int64_value
+uint64 uint64_value
+string string_value
+---
+bool bool_value
+byte byte_value
+char char_value
+float32 float32_value
+float64 float64_value
+int8 int8_value
+uint8 uint8_value
+int16 int16_value
+uint16 uint16_value
+int32 int32_value
+uint32 uint32_value
+int64 int64_value
+uint64 uint64_value
+string string_value
