+++ msg-gen-macro/test_msgs/srv/Empty.srv
@@ -1,0 +1,1 @@
+---
