+++ msg-gen/src/lib.rs
@@ -10,38 +10,4 @@
 mod parser;
 mod types;
 
-use std::path::Path;
-
-use proc_macro::TokenStream;
-use quote::quote;
-
-use crate::parser::get_packages;
-
-#[proc_macro]
-pub fn msg_include_all(_input: TokenStream) -> TokenStream {
-    let ament_prefix_path = std::env!("DETECTED_AMENT_PREFIX_PATH").trim();
-    if ament_prefix_path.is_empty() {
-        quote! {
-            /// **No messages are available because the `AMENT_PREFIX_PATH` environment variable
-            /// was not set during build.**
-            pub const AMENT_PREFIX_PATH_NOT_SET: () = ();
-        }
-        .into()
-    } else {
-        let paths = ament_prefix_path
-            .split(':')
-            .map(Path::new)
-            .collect::<Vec<_>>();
-
-        let packages = get_packages(&paths)
-            .unwrap()
-            .iter()
-            .map(|v| v.token_stream())
-            .collect::<Vec<_>>();
-
-        (quote! {
-            #(#packages)*
-        })
-        .into()
-    }
-}
+pub use crate::parser::get_packages;
