+++ src/lib.rs
@@ -1,6 +1,6 @@
 pub use ros2_client;
 pub use rustdds;
 
-dora_ros2_bridge_msg_gen::msg_include_all!();
+dora_ros2_bridge_msg_gen_macro::msg_include_all!();
 
 pub mod _core;
